+++ src/main/java/com/springkafka/kafka_app/service/kafka_streams/KafkaStreamsService.java
@@ -6,6 +6,7 @@ import com.springkafka.kafka_app.utils.Query.AttributeType;
 import com.springkafka.kafka_app.utils.Query.Query;
 import com.springkafka.kafka_app.utils.ServiceProperties;
 import com.springkafka.kafka_app.utils.TopicEnum;
+import com.springkafka.kafka_app.utils.serdes.EventSerde;
 import com.springkafka.kafka_app.utils.serdes.HashMapSerde;
 import com.springkafka.kafka_app.wrapper.CustomLogger;
 import org.apache.kafka.common.serialization.Serdes;
@@ -13,6 +14,7 @@ import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.streams.KafkaStreams;
 import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.StreamsBuilder;
+import org.apache.kafka.streams.StreamsConfig;
 import org.apache.kafka.streams.kstream.KStream;
 import org.apache.kafka.streams.kstream.KTable;
 import org.apache.kafka.streams.kstream.Materialized;
@@ -27,35 +29,40 @@ import java.util.Properties;
 
 @Service
 public class KafkaStreamsService extends CustomLogger {
-    private final StreamsKafka streamsKafka;
     private KafkaStreams kafkaStreams;
 
     @Autowired
-    public KafkaStreamsService(StreamsKafka streamsKafka) {
-        this.streamsKafka = streamsKafka;
+    public KafkaStreamsService() {
         Runtime.getRuntime().addShutdownHook(new Thread(this::shutdown));
+    }
+    public Properties getProperties(String topic){
+        Properties properties = new Properties();
+        properties.put(StreamsConfig.APPLICATION_ID_CONFIG, topic);
+        properties.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, ServiceProperties.KAFKA_BROKERS);
+        properties.put(StreamsConfig.DEFAULT_KEY_SERDE_CLASS_CONFIG, Serdes.String().getClass());
+        properties.put(StreamsConfig.DEFAULT_VALUE_SERDE_CLASS_CONFIG, EventSerde.class);
+        properties.put(StreamsConfig.CACHE_MAX_BYTES_BUFFERING_CONFIG,0);
+        return properties;
     }
 
 
 
     public void getFilteredStream(Query query, String topic){
-        StreamsBuilder streamsBuilder = streamsKafka.getStreamsBuilder();
-        Properties config = streamsKafka.getProperties(topic);
+        StreamsBuilder streamsBuilder = new StreamsBuilder();
+        Properties config = getProperties(topic);
 
-//        StoreBuilder<KeyValueStore<String, Long>> keyValueStoreStoreBuilder = Stores.keyValueStoreBuilder
-//                        (Stores.inMemoryKeyValueStore(ServiceProperties.USER_COUNT_STORE),
-//                                Serdes.String(),
-//                                Serdes.Long())
-//                .withLoggingDisabled();
-//
-//        streamsBuilder.addStateStore(keyValueStoreStoreBuilder);
-
-//        long startTime = query.getTimestamp().getStartTime();
-//        long endTime = query.getTimestamp().getEndTime();
+        long startTime = query.getTimestamp().getStartTime();
+        long endTime = query.getTimestamp().getEndTime();
 
         KStream<String, Event> inputStream = streamsBuilder.stream(TopicEnum.TOPIC.getTopicName());
 
-        KTable<String, Map<String, Integer>> userAttributeCountTable = inputStream
+        KStream<String ,Event> timeFilterStream = inputStream
+                .filter((key,event) -> {
+                    long eventTime = Long.valueOf(event.getMapKeyValue("eventTime").toString());
+                    return eventTime>=startTime && eventTime<=endTime;
+                });
+
+        KTable<String, Map<String, Integer>> userAttributeCountTable = timeFilterStream
                 .groupBy((key,event) -> event.getMapKeyValue("name").toString())
                 .aggregate(
                         HashMap::new,
@@ -79,6 +86,7 @@ public class KafkaStreamsService extends CustomLogger {
                         for(Attribute attribute : attributeType.getAttributeList()){
                             String attributeValue = attribute.getValue();
                             Integer count = attribute.getCount();
+                            info("the count given is {} and the gotten count is {}",count,attributeCount.getOrDefault(attributeValue,0));
                             if(attributeCount.getOrDefault(attributeValue,0)<count) {
                                 return false;
                             }
@@ -86,7 +94,6 @@ public class KafkaStreamsService extends CustomLogger {
                     }
                     return true;
                 });
-
 
         KStream<String,String> outputStream = userAttributeCountTable
                 .toStream()
